Move heuristics to a separate package
